学生 查询在线考试列表

+++ src/main/java/com/twp/controller/StuInfoController.java
@@ -96,7 +96,7 @@ public class StuInfoController {
 	//@RequiresPermissions("stuinfo:info")
 	public R queryStuClass(){
 		Long userId = ShiroUtils.getUserId();
-		Long classId = stuInfoService.queryStuClass(userId);
+		int classId = stuInfoService.queryStuClass(userId);
 		return R.ok().put("classId", classId);
 	}
