+++ src/main/java/com/twp/service/StuInfoService.java
@@ -18,7 +18,7 @@ public interface StuInfoService {
 	
 	List<StuInfoEntity> queryList(Map<String, Object> map);
 	
-	Long queryStuClass(long userId);
+	int queryStuClass(long userId);
 	
 	int queryTotal(Map<String, Object> map);
